feature: adicionando e removendo itens do carrinho
