+++ src/components/card/style.module.css
@@ -32,7 +32,6 @@
 
   height: 2.875rem;
   color: var(--white);
-  background: var(--light-gray);
   font-size: 1.5rem;
 }
 
@@ -49,21 +48,11 @@
 }
 
 .add-item-button {
-  width: 3.5rem;
+  width: 100%;
   color: var(--white);
   background: var(--medium-gray);
   border: none;
   outline: none;
-  font-size: 2rem;
-  border-radius: 0 0 1rem 0;
-}
-
-.remove-item-button {
-  width: 3.5rem;
-  color: var(--white);
-  background: var(--medium-gray);
-  border: none;
-  outline: none;
-  font-size: 2rem;
-  border-radius: 0 0 0 1rem;
+  font-size: 1rem;
+  border-radius: 0 0 1rem 1rem;
 }
