+++ src/components/header/header.css
@@ -1,0 +1,91 @@
+.navbar {
+  display: flex;
+  justify-content: space-around;
+  align-items: center;
+  height: 75px;
+  background-color: var(--dark-gray);
+  color: var(--neon-green);
+  z-index: 1;
+}
+
+.logo {
+  font-size: 30px;
+}
+
+.nav-links {
+  display: flex;
+  justify-content: flex-end;
+  list-style: none;
+  width: 75%;
+}
+
+.menu,
+.pedidos,
+.logout {
+  text-decoration: none;
+  color: var(--neon-green);
+  font-size: 1.5rem;
+  padding: 10px 20px;
+  margin: 0 5px;
+  cursor: pointer;
+}
+
+.logout {
+  border: 1px solid var(--neon-green);
+  border-radius: 5px;
+}
+
+.mobile-menu-icon {
+  display: none;
+}
+
+@media screen and (max-width: 780px) {
+  .logo {
+    display: flex;
+    position: absolute;
+    top: 15px;
+    left: 35px;
+  }
+  .nav-links {
+    visibility: hidden;
+    height: 0;
+    width: 100vw;
+  }
+  .nav-links-mobile {
+    position: absolute;
+    display: block;
+    visibility: visible;
+    overflow-y: hidden;
+    height: calc(100vh - 840px);
+    list-style: none;
+    background-color: var(--dark-gray);
+    top: 60px;
+    transition: 0.2s;
+    border-radius: 0 0 20px 20px;
+    width: 100vw;
+    z-index: 1;
+  }
+  .menu,
+  .pedidos,
+  .logout {
+    color: var(--neon-green);
+    text-align: center;
+    padding: 32px;
+    width: 100%;
+    text-transform: uppercase;
+  }
+  .logout {
+    border: none;
+  }
+  .mobile-menu-icon {
+    display: block;
+    position: absolute;
+    font-size: 30px;
+    color: var(--neon-green);
+    background-color: var(--dark-gray);
+    border: none;
+    outline: none;
+    top: 15px;
+    right: 25px;
+  }
+}
