+++ src/components/header/index.js
@@ -1,6 +1,6 @@
 import { React, useState } from 'react';
 import { Link } from 'react-router-dom';
-import './menu.css';
+import './header.css';
 
 export default function Header() {
   const [isMobile, setIsMobile] = useState(false);
