--- src/components/header/menu.css
@@ -1,244 +1,0 @@
-.navbar {
-  display: flex;
-  justify-content: space-around;
-  align-items: center;
-  height: 75px;
-  background-color: var(--dark-gray);
-  color: var(--neon-green);
-  z-index: 1;
-}
-
-.logo {
-  font-size: 30px;
-}
-
-.nav-links {
-  display: flex;
-  justify-content: flex-end;
-  list-style: none;
-  width: 75%;
-}
-
-.menu,
-.pedidos,
-.logout {
-  text-decoration: none;
-  color: var(--neon-green);
-  font-size: 1.5rem;
-  padding: 10px 20px;
-  margin: 0 5px;
-  cursor: pointer;
-}
-
-.logout {
-  border: 1px solid var(--neon-green);
-  border-radius: 5px;
-}
-
-.mobile-menu-icon {
-  display: none;
-}
-
-@media screen and (max-width: 780px) {
-  .logo {
-    display: flex;
-    position: absolute;
-    top: 15px;
-    left: 35px;
-  }
-  .nav-links {
-    visibility: hidden;
-    height: 0;
-    width: 100vw;
-  }
-  .nav-links-mobile {
-    position: absolute;
-    display: block;
-    visibility: visible;
-    overflow-y: hidden;
-    height: calc(100vh - 840px);
-    list-style: none;
-    background-color: var(--dark-gray);
-    top: 60px;
-    transition: 0.2s;
-    border-radius: 0 0 20px 20px;
-    width: 100vw;
-    z-index: 1;
-  }
-  .menu,
-  .pedidos,
-  .logout {
-    color: var(--neon-green);
-    text-align: center;
-    padding: 32px;
-    width: 100%;
-    text-transform: uppercase;
-  }
-  .logout {
-    border: none;
-  }
-  .mobile-menu-icon {
-    display: block;
-    position: absolute;
-    font-size: 30px;
-    color: var(--neon-green);
-    background-color: var(--dark-gray);
-    border: none;
-    outline: none;
-    top: 15px;
-    right: 25px;
-  }
-}
-
-/*orders*/
-
-.footer {
-  position: fixed;
-  display: flex;
-  align-items: center;
-  flex-direction: column;
-  top: 68rem;
-  width: 100vw;
-  background: var(--dark-gray);
-}
-
-.subtitles-container {
-  display: grid;
-  grid-template-columns: 20rem 8rem 11rem 1fr;
-  gap: 2rem;
-  text-align: center;
-  justify-content: space-around;
-  color: var(--white);
-  font-size: 2rem;
-  margin: 2rem;
-  padding: 1rem;
-  border-bottom: 1px solid var(--light-gray);
-}
-
-.item-subtitle {
-  text-align: start;
-}
-
-.item-list {
-  display: grid;
-  grid-template-columns: 24rem 3rem 12rem 5rem;
-  gap: 4rem;
-  color: var(--white);
-  justify-content: space-around;
-  align-items: center;
-  font-size: 2rem;
-}
-
-.itens {
-  display: flex;
-  justify-content: space-around;
-  align-items: center;
-  height: 36px;
-  color: var(--white);
-}
-
-.resultOrders {
-  display: flex;
-  color: var(--white);
-  bottom: 5px;
-  width: 100vw;
-}
-
-.btn-cancel {
-  width: 17rem;
-  height: 3.75rem;
-  border-radius: 2.625rem;
-  background: #f25764;
-  font-size: 2rem;
-  border: none;
-  color: white;
-  font-weight: bold;
-}
-.btn-confirm {
-  width: 20rem;
-  height: 3.75rem;
-  border-radius: 2.625rem;
-  background: #0dd99f;
-  font-size: 2rem;
-  border: none;
-  color: white;
-  font-weight: bold;
-}
-
-.close-orders {
-  visibility: hidden;
-  height: 0;
-  width: 100vw;
-}
-
-.open-orders {
-  position: absolute;
-  visibility: visible;
-  overflow-y: hidden;
-  height: calc(100vh - 24rem);
-  list-style: none;
-  background-color: var(--dark-gray);
-  bottom: 5rem;
-  left: 0;
-  border-radius: 20px 20px 0 0;
-  transition: 0.4s;
-  width: 100vw;
-}
-
-.btn-orders {
-  color: var(--neon-green);
-  background: none;
-  font-size: 30px;
-  display: flex;
-  height: 75px;
-  width: 100vw;
-  align-items: center;
-  justify-content: center;
-}
-
-.buttons {
-  position: absolute;
-  top: 44rem;
-  display: flex;
-  justify-content: space-evenly;
-  width: 100vw;
-}
-
-.total-price {
-  position: absolute;
-  font-size: 2rem;
-  top: 40rem;
-  display: flex;
-  justify-content: space-around;
-  width: 100vw;
-}
-
-.trash {
-  background: none;
-  border: none;
-  text-align: start;
-}
-
-.fa-trash-alt {
-  font-size: 2.5rem;
-  color: var(--dark-pink);
-}
-
-.fa-hamburger {
-  margin: 1rem;
-}
-
-.item-number {
-  background: var(--background2);
-  padding: 14px 12px 12px 12px;
-  border-radius: 8px;
-  border: var(--light-gray) 1px solid;
-}
-
-.item-price {
-  text-align: center;
-}
-
-.item-name {
-  text-align: center;
-}
