+++ src/pages/register/style.module.css
@@ -3,3 +3,128 @@
   justify-content: center;
   margin: 4rem;
 }
+
+.box {
+  background: white;
+  width: 30rem;
+  height: 26.75rem;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+
+  border-radius: 3rem;
+}
+
+.subTitle {
+  font-size: 3rem;
+  color: var(--medium-purple);
+  text-align: center;
+  margin: 2rem;
+  text-transform: uppercase;
+  font-weight: 600;
+}
+
+.inputBox {
+  width: 25.125rem;
+  height: 3.5rem;
+  margin: 0.6rem;
+  outline-color: var(--medium-purple);
+  border: 1px solid var(--light-purple);
+  font-size: 1.5rem;
+  text-align: center;
+
+  border-radius: 9px;
+}
+
+.inputBox::placeholder {
+  color: var(--light-purple);
+}
+
+.link-login {
+  text-decoration: none;
+  color: var(--neon-green);
+  font-size: 1.5rem;
+  margin-top: 0.5rem;
+}
+
+.form {
+  background: var(--white);
+  width: 32rem;
+  height: 48.5rem;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  margin: auto;
+  border-radius: 3rem;
+  margin-top: 3rem;
+}
+
+.error {
+  text-align: center;
+  color: var(--dark-pink);
+  height: 0.9rem;
+  width: 100%;
+  font-size: 15px;
+  margin-bottom: 0.3rem;
+}
+
+.register {
+  font-size: 2.9rem;
+  color: var(--neon-green);
+  text-align: center;
+  margin: 2rem;
+}
+
+.form-input {
+  width: 25.125rem;
+  height: 3.5rem;
+  margin: 0.5rem;
+  border: 1px solid var(--neon-green);
+  outline-color: var(--neon-green);
+  font-size: 1.5rem;
+  text-align: center;
+  border-radius: 9px;
+}
+
+.form-input::placeholder {
+  color: var(--neon-green);
+}
+
+.btn-register {
+  margin-top: 2rem;
+  width: 21.875rem;
+  background: var(--neon-green);
+  height: 3.75rem;
+  font-size: 1.5rem;
+  text-transform: uppercase;
+  border-radius: 2.5rem;
+  color: white;
+  border: none;
+  font-weight: bold;
+}
+
+.select {
+  width: 25rem;
+  height: 3.5rem;
+  padding: 8px 12px;
+  margin: 0.6rem;
+  color: var(--neon-green);
+  background: var(--white);
+  border: 1px solid var(--neon-green);
+  cursor: pointer;
+  border-radius: 9px;
+  font-size: 1.5rem;
+  text-align: center;
+}
+
+.span-register {
+  text-decoration: none;
+  color: var(--neon-green);
+  font-size: 1.5rem;
+  margin-top: 1.5rem;
+}
+
+.success {
+  font-size: 2rem;
+  text-align: center;
+}
